+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - The 'printfbroadcastinfo()' function is provided to query the broadcast information in the current system to help developers clean up useless broadcasts in time
 
+- Avoid error callback calls in two-way communication
+
 ## 1.0.2
 
 - StickyBroadcast supports multiple messages of the same type in the broadcast system at the same time